+++ hydrofunctions/__init__.py
@@ -98,4 +98,5 @@ from .usgs_rdb import (
         read_rdb,
         field_meas,
         rating_curve,
+        peaks,
         )